+++ src/main/java/com/buixuantruong/shopapp/model/OrderDetail.java
@@ -32,7 +32,7 @@ public class OrderDetail {
     private Integer numberOfProducts;
 
     @Column(name = "total_money", nullable = false)
-    private Integer totalMoney;
+    private Long totalMoney;
 
     @Column(name = "color")
     private String color;
